FIX: circular import and numpy docs
