+++ regolith/publistbuilder.py
@@ -14,8 +14,8 @@ try:
 except ImportError:
     HAVE_BIBTEX_PARSER = False
 
-from regolith.tools import all_docs_from_collection, date_to_float, \
-    date_to_rfc822, rfc822now, gets, month_and_year
+from regolith.tools import all_docs_from_collection, date_to_rfc822, rfc822now, gets, month_and_year
+from regolith.dates import date_to_float
 from regolith.sorters import doc_date_key, ene_date_key, category_val, \
     level_val, id_key, date_key, position_key
