+++ regolith/tools.py
@@ -1,19 +1,20 @@
 """Misc. regolith tools.
 """
-import os
-import re
-import sys
-import platform
 import email.utils
+import os
+import platform
+import sys
 from copy import deepcopy
 
 from datetime import datetime
 
+from regolith.dates import month_to_int, date_to_float
 from regolith.sorters import doc_date_key, id_key, ene_date_key
 
 try:
     from bibtexparser.bwriter import BibTexWriter
     from bibtexparser.bibdatabase import BibDatabase
+
     HAVE_BIBTEX_PARSER = True
 except ImportError:
     HAVE_BIBTEX_PARSER = False
@@ -52,8 +53,10 @@ def dbpathname(db, rc):
 def fallback(cond, backup):
     """Decorator for returning the object if cond is true and a backup if cond is false.
     """
+
     def dec(obj):
         return obj if cond else backup
+
     return dec
 
 
@@ -67,65 +70,8 @@ def all_docs_from_collection(client, collname):
         yield from client.all_documents(dbname, collname)
 
 
-MONTHS = {
-    'jan': 1,
-    'jan.': 1,
-    'january': 1,
-    'feb': 2,
-    'feb.': 2,
-    'february': 2,
-    'mar': 3,
-    'mar.': 3,
-    'march': 3,
-    'apr': 4,
-    'apr.': 4,
-    'april': 4,
-    'may': 5,
-    'may.': 5,
-    'jun': 6,
-    'jun.': 6,
-    'june': 6,
-    'jul': 7,
-    'jul.': 7,
-    'july': 7,
-    'aug': 8,
-    'aug.': 8,
-    'august': 8,
-    'sep': 9,
-    'sep.': 9,
-    'sept': 9,
-    'sept.': 9,
-    'september': 9,
-    'oct': 10,
-    'oct.': 10,
-    'october': 10,
-    'nov': 11,
-    'nov.': 11,
-    'november': 11,
-    'dec': 12,
-    'dec.': 12,
-    'december': 12,
-    '': 1
-    }
-
 SHORT_MONTH_NAMES = (None, 'Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul',
                      'Aug', 'Sept', 'Oct', 'Nov', 'Dec')
-
-def month_to_int(m):
-    """Converts a month to an integer."""
-    try:
-        m = int(m)
-    except ValueError:
-        m = MONTHS[m.lower()]
-    return m
-
-
-def date_to_float(y, m, d=0):
-    """Converts years / months / days to a float, eg 2015.0818 is August 18th 2015."""
-    y = int(y)
-    m = month_to_int(m)
-    d = int(d)
-    return y + (m/100.0) + (d/100000.0)
 
 
 def date_to_rfc822(y, m, d=1):
